# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gym API",
        "description": "API pour la gestion de salle de sport",
        "contact": {
            "email": "admin@gym-gestion.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/rooms/new": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Ajouter une nouvelle salle",
                "operationId": "a726f9ee5a345991beaf1e89765f4455",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "capacity"
                                ],
                                "properties": {
                                    "room_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Salle A"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "length": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20.5
                                    },
                                    "width": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "height": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Création avec succès"
                    },
                    "400": {
                        "description": "Demande invalide"
                    }
                }
            }
        },
        "/api/rooms/status/{id}": {
            "patch": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Activer/Désactiver une salle",
                "operationId": "6bb8b0a74b612cf709da98f4d5fd5f28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réussite"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                }
            }
        },
        "/api/rooms/update/{id}": {
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Mettre les informations relatives à une salle",
                "operationId": "017102629a3e515381f7b470ea93e881",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "room_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Salle A"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "length": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20.5
                                    },
                                    "width": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "height": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie"
                    },
                    "404": {
                        "description": "Salle introuvable"
                    }
                }
            }
        },
        "/api/rooms/delete/{id}": {
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Suppression d'une salle",
                "operationId": "ef047618b43ac3f380c2e1d99274bb09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression effectué avec succès"
                    },
                    "401": {
                        "description": "Vous n'êtes pas autorisé à réaliser cette action"
                    },
                    "404": {
                        "description": "Salle introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}